updated the hw_avg count (set_avgs) form 18bit to 32 bit counter
This should significantly extend the limit for how many experiments one
can do in streaming mode (which can in principle be unlimited with more
work).    Note however: the 18bit was set to ensure that there was no
overflowing of the accumulators in hw averaging mode, so it is now
possible to do too many averages.  However this also allows you to do
more averages for data that you may know does not use all of the bits.
So some good but also some lack of safety.